scodec
Migration to scodec

Src is compiling

tests compiling

All unit tests passing now

fix compiler issue in ScriptParser

bump timeout to run on travis ci

Fixing two unit tests that had types failing

Turn down excessive logging

Uncomment rpc tests

fixing nits

+++ core/src/main/scala/org/bitcoins/core/script/constant/BytesToPushOntoStack.scala
@@ -25,7 +25,7 @@ object BytesToPushOntoStack extends ScriptOperationFactory[BytesToPushOntoStack]
   override def operations: Seq[BytesToPushOntoStack] =
     (for { i <- 0 to 75 } yield BytesToPushOntoStackImpl(i))
 
-  def fromNumber(num: Int): BytesToPushOntoStack = {
+  def fromNumber(num: Long): BytesToPushOntoStack = {
     if (num > 75) throw new IllegalArgumentException("We cannot have a BytesToPushOntoStack for greater than 75 bytes")
     else {
       val bytesToPushOntoStackOpt = operations.find(_.opCode == num)
@@ -36,5 +36,5 @@ object BytesToPushOntoStack extends ScriptOperationFactory[BytesToPushOntoStack]
     }
   }
 
-  def apply(num: Int): BytesToPushOntoStack = fromNumber(num)
+  def apply(num: Long): BytesToPushOntoStack = fromNumber(num)
 }
